Issue-#12-Scrapper/DB Changes
Need to also make sure these changes are all reflected in accessors

- Renamed pokemonID to uniquePokemonID to better represent that it refers to each individual instance of a different pokemon form
- Refactored pokemonID->uniquePokemonID where ever used
- Renamed pokemon table to pokemon_unique_info
- Combined caught, description, hatchTime, catchRate, genderRatioMale attributes from tables into one pokemon_common_info table
- Changed use of uniquePokemonID (formerly pokemonID) in pokemon_eggGroups, pokemon_locations and pokemon_moves, pokemon_caught to use nationalID instead
- Tables Deleted: pokemon_description, pokemon_catchRate, pokemon_genderRatioMale, pokemon_hatchTime

-Changed scrapper_description  to reflect the above changes and re-ran it to update

-Added a check in move_database to check if .db file is present before removing

+++ Scrapper/move_database.py
@@ -11,7 +11,8 @@ here = "../database/pokedex.sqlite3"
 there = "../app/src/main/assets/pokedex.db"
 
 # first remove the old database
-os.remove(there)
+if os.path.isfile(there):
+    os.remove(there)
 
 # then copy the updated database there
 shutil.copyfile(here, there)
